refactor: move chat social access owner
